made it work

+++ scripts/start_dev_db.sh
@@ -1,15 +1,24 @@
-# Run with `source`
+# This script is meant to be run with `source` so it can set environment variables.
 
 export PGDATA="$PWD/dev_pgdata"
-export LEMMY_DATABASE_URL="postgresql:///lemmy?host=$PWD"
+export PGHOST=$PWD
+export LEMMY_DATABASE_URL="postgresql://lemmy:password@/lemmy?host=$PWD"
 
-rm -rf $PGDATA
-initdb --username=lemmy --auth=trust
+# If cluster exists, stop the server and delete the cluster
+if [ -d $PGDATA ]
+then
+  # Prevent `stop` from failing if server already stopped
+  pg_ctl restart > /dev/null
+  pg_ctl stop
+  rm -rf $PGDATA
+fi
 
-postgres -c listen_addresses= -c unix_socket_directories=$PWD
+# Create cluster
+initdb --username=postgres --auth=trust --no-instructions
 
-# Wait for socket to exist
-while test "$(echo *.PGSQL.*)" == '*.PGSQL.*'
-do
-  sleep 0.1
-done
+# Start server that only listens to socket in current directory
+pg_ctl start --options="-c listen_addresses= -c unix_socket_directories=$PWD" > /dev/null
+
+# Setup database
+psql -c "CREATE USER lemmy WITH PASSWORD 'password' SUPERUSER;" -U postgres
+psql -c "CREATE DATABASE lemmy WITH OWNER lemmy;" -U postgres
